ml-service/requirements.txt
# ML Microservice Requirements
# Deploy separately on Railway.app (free tier has 8GB memory)

flask==3.0.0
flask-cors==4.0.0
gunicorn==21.2.0

# Data Processing
pandas==2.2.3
numpy==1.26.4

# Machine Learning
scikit-learn==1.5.2

# Resume Parsing
PyPDF2==3.0.1
python-docx==1.1.0

# NLP (lightweight)
spacy==3.7.2

# API
requests==2.32.3
python-dotenv==1.0.0
